requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
PyJWT==2.4.0
web3==5.29.0